+++ app/mailers/admin_mailer.rb
@@ -1,45 +1,45 @@
 class AdminMailer < ApplicationMailer
   helper :messages
 
-  def new_developer_profile
-    @notification = params[:record]
+  def new_developer
+    @notification = params[:record].to_notification
     recipient = params[:recipient]
 
-    @developer = @notification.to_notification.developer
+    @developer = @notification.developer
 
-    mail(to: recipient.email, subject: "New developer profile added")
+    mail(to: recipient.email, subject: @notification.email_subject)
   end
 
   def new_business
-    @notification = params[:record]
+    @notification = params[:record].to_notification
     recipient = params[:recipient]
 
-    @business = @notification.to_notification.business
+    @business = @notification.business
 
-    mail(to: recipient.email, subject: "New business added")
+    mail(to: recipient.email, subject: @notification.email_subject)
   end
 
   def new_conversation
-    @notification = params[:record]
+    @notification = params[:record].to_notification
     recipient = params[:recipient]
 
-    conversation = @notification.to_notification.conversation
+    conversation = @notification.conversation
     @business = conversation.business
     @developer = conversation.developer
     @subscriptions = @business.user.subscriptions
     @body = conversation.messages.first.body
 
-    mail(to: recipient.email, subject: "New conversation started")
+    mail(to: recipient.email, subject: @notification.email_subject)
   end
 
   def potential_hire
-    @notification = params[:record]
+    @notification = params[:record].to_notification
     recipient = params[:recipient]
 
-    @developer = @notification.to_notification.developer
+    @developer = @notification.developer
     @conversations = @developer.conversations.count
     @replies = @developer.messages.distinct.count(:conversation_id)
 
-    mail(to: recipient.email, subject: "New potential hire")
+    mail(to: recipient.email, subject: @notification.email_subject)
   end
 end