+++ db/migrate/20220519230743_remove_email_from_developers.rb
@@ -1,0 +1,5 @@
+class RemoveEmailFromDevelopers < ActiveRecord::Migration[7.0]
+  def change
+    remove_column :developers, :email, :string
+  end
+end