+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_05_01_021745) do
+ActiveRecord::Schema[7.0].define(version: 2022_05_23_115740) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -77,7 +77,6 @@ ActiveRecord::Schema[7.0].define(version: 2022_05_01_021745) do
 
   create_table "developers", force: :cascade do |t|
     t.string "name", null: false
-    t.string "email"
     t.date "available_on"
     t.string "hero", null: false
     t.text "bio", null: false
@@ -95,6 +94,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_05_01_021745) do
     t.integer "preferred_max_salary"
     t.virtual "textsearchable_index_col", type: :tsvector, as: "to_tsvector('simple'::regconfig, (((COALESCE(hero, ''::character varying))::text || ' '::text) || COALESCE(bio, ''::text)))", stored: true
     t.datetime "featured_at"
+    t.boolean "profile_reminder_notifications", default: true
     t.index ["textsearchable_index_col"], name: "textsearchable_index", using: :gin
     t.index ["user_id"], name: "index_developers_on_user_id"
   end