+++ test/integration/stripe_checkout_test.rb
@@ -7,21 +7,13 @@ class StripeCheckoutTest < ActionDispatch::IntegrationTest
     @user = users(:business)
   end
 
-  test "redirects to the Stripe Checkout URL" do
-    sign_in @user
-
-    stub_pay(@user) do
-      post stripe_checkout_path
-      assert_redirected_to "checkout.stripe.com"
-    end
-  end
-
-  test "passes along the plan" do
+  test "passes along the plan and redirects to the Stripe Checkout URL" do
     sign_in @user
     full_time_price_id = Rails.application.credentials.stripe[:price_ids][:full_time_plan]
 
     stub_pay(@user, plan_price_id: full_time_price_id) do
       post stripe_checkout_path(plan: :full_time)
+      assert_redirected_to "checkout.stripe.com"
     end
   end
 
@@ -31,7 +23,7 @@ class StripeCheckoutTest < ActionDispatch::IntegrationTest
     sign_in @user
 
     stub_pay(@user) do
-      post stripe_checkout_path
+      post stripe_checkout_path(plan: :part_time)
       assert_equal Analytics::Event.last.url, success_path
     end
   end